requirements.txt
Flask==3.1.2
pymongo==4.15.1
python-dotenv==1.1.1
Flask-PyMongo==3.0.1
flask-pydantic==0.13.2
